+++ web-view/viewer/src/jsMain/kotlin/com/firsttimeinforever/intellij/pdf/viewer/application/pdfjs/Internals.kt
@@ -1,0 +1,14 @@
+package com.firsttimeinforever.intellij.pdf.viewer.application.pdfjs
+
+internal object Internals {
+  object StyleVariables {
+    const val mainColor = "--main-color"
+    const val bodyBackgroundColor = "--body-bg-color"
+    const val sidebarNarrowBackgroundColor = "--sidebar-narrow-bg-color"
+    const val iconsBackgroundColor = "--toolbar-icon-bg-color"
+    const val treeItemColor = "--treeitem-color"
+    const val treeItemHoverColor = "--treeitem-hover-color"
+    const val treeItemSelectedColor = "--treeitem-selected-color"
+    const val treeItemSelectedBackgroundColor = "--treeitem-selected-bg-color"
+  }
+}