+++ web-view/viewer/src/jsMain/kotlin/com/firsttimeinforever/intellij/pdf/viewer/application/pdfjs/types/PdfLinkService.kt
@@ -1,0 +1,5 @@
+package com.firsttimeinforever.intellij.pdf.viewer.application.pdfjs.types
+
+external class PdfLinkService {
+  fun goToDestination(destination: dynamic)
+}