+++ web-view/viewer/src/main/kotlin/com/firsttimeinforever/intellij/pdf/viewer/application/pdfjs/types/PdfLinkService.kt
@@ -1,5 +1,0 @@
-package com.firsttimeinforever.intellij.pdf.viewer.application.pdfjs.types
-
-external class PdfLinkService {
-  fun navigateTo(destination: String)
-}